DF-11514 testbed: for now DeviceManagerTest only for windows
Former-commit-id: 1aa4c4ee5dbc7a05c8cb963491b48366879822db

+++ Projects/TestBed/Classes/Tests/DeviceManagerTest.cpp
@@ -126,7 +126,7 @@ void DeviceManagerTest::OnDisplayClick(DAVA::BaseObject*, void* args, void*)
 {
     const Vector<DisplayInfo>& displays = deviceManager->GetDisplays();
     int i = static_cast<int>(reinterpret_cast<intptr_t>(args));
-    if (0 <= i && i < displays.size())
+    if (0 <= i && i < static_cast<int>(displays.size()))
     {
         StringStream ss;
         ss << displays[i].name;
